Split backend sync contract tests
